# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Al Hiraa'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
  /api/v1/request-otp:
    post:
      summary: 'Request OTP'
      operationId: requestOTP
      description: 'This endpoint will be used to request OTP for login.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'OTP sent successfully.'
                properties:
                  message:
                    type: string
                    example: 'OTP sent successfully.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile_number:
                  type: integer
                  description: 'Mobile Number of the user.'
                  example: 9876543210
              required:
                - mobile_number
      security: []
...